datadog-crawler/requirements.txt
requests
beautifulsoup4
feedparser
schedule
fastapi
uvicorn
pymongo
apscheduler
python-dotenv
python-dateutil
confluent_kafka
openai
ddtrace
# 키워드 분석 및 시각화용 라이브러리
konlpy
scikit-learn
wordcloud
plotly
networkx
matplotlib
seaborn
pandas
numpy
collections-extended